Reorganized grapes module
