+++ src/java/growthcraft/grapes/ClientProxy.java
@@ -1,6 +1,10 @@
 package growthcraft.grapes;
 
 import growthcraft.core.GrowthCraftCore;
+import growthcraft.grapes.renderer.RenderGrape;
+import growthcraft.grapes.renderer.RenderGrapeLeaves;
+import growthcraft.grapes.renderer.RenderGrapeVine1;
+
 import cpw.mods.fml.client.registry.RenderingRegistry;
 
 public class ClientProxy extends CommonProxy
