+++ src/java/growthcraft/grapes/block/BlockGrapeBlock.java
@@ -1,7 +1,12 @@
-package growthcraft.grapes;
+package growthcraft.grapes.block;
 
 import java.util.Random;
 
+import growthcraft.grapes.GrowthCraftGrapes;
+import growthcraft.grapes.renderer.RenderGrape;
+
+import cpw.mods.fml.relauncher.Side;
+import cpw.mods.fml.relauncher.SideOnly;
 import net.minecraft.block.Block;
 import net.minecraft.block.material.Material;
 import net.minecraft.client.renderer.texture.IIconRegister;
@@ -11,8 +16,6 @@ import net.minecraft.item.Item;
 import net.minecraft.util.AxisAlignedBB;
 import net.minecraft.util.IIcon;
 import net.minecraft.world.World;
-import cpw.mods.fml.relauncher.Side;
-import cpw.mods.fml.relauncher.SideOnly;
 
 public class BlockGrapeBlock extends Block
 {
