+++ src/java/growthcraft/grapes/block/BlockGrapeVine0.java
@@ -1,20 +1,17 @@
-package growthcraft.grapes;
+package growthcraft.grapes.block;
 
-import growthcraft.core.GrowthCraftCore;
-
-import java.util.List;
 import java.util.Random;
 
-import cpw.mods.fml.common.eventhandler.Event.Result;
+import growthcraft.grapes.GrowthCraftGrapes;
+import growthcraft.core.integration.AppleCore;
+
 import cpw.mods.fml.common.eventhandler.Event;
+import cpw.mods.fml.common.eventhandler.Event.Result;
 import cpw.mods.fml.relauncher.Side;
 import cpw.mods.fml.relauncher.SideOnly;
-import growthcraft.core.integration.AppleCore;
 import net.minecraft.block.Block;
-import net.minecraft.block.BlockLeaves;
 import net.minecraft.block.material.Material;
 import net.minecraft.client.renderer.texture.IIconRegister;
-import net.minecraft.entity.Entity;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.init.Blocks;
 import net.minecraft.item.Item;
@@ -26,22 +23,20 @@ import net.minecraftforge.common.EnumPlantType;
 import net.minecraftforge.common.IPlantable;
 import net.minecraftforge.common.util.ForgeDirection;
 
-public class BlockGrapeVine1 extends Block implements IPlantable
+public class BlockGrapeVine0 extends Block implements IPlantable
 {
-	private final float growth = GrowthCraftGrapes.grapeVine1_growth;
+	private final float growth = GrowthCraftGrapes.grapeVine0_growth;
 
 	@SideOnly(Side.CLIENT)
 	public static IIcon[] tex;
-	public boolean graphicFlag;
 
-	public BlockGrapeVine1()
+	public BlockGrapeVine0()
 	{
 		super(Material.plants);
 		this.setTickRandomly(true);
-		this.setHardness(2.0F);
-		this.setResistance(5.0F);
-		this.setStepSound(soundTypeWood);
-		this.setBlockName("grc.grapeVine1");
+		this.setHardness(0.0F);
+		this.setStepSound(soundTypeGrass);
+		this.setBlockName("grc.grapeVine0");
 		this.setCreativeTab(null);
 	}
 
@@ -49,7 +44,7 @@ public class BlockGrapeVine1 extends Block implements IPlantable
 	{
 		int previousMetadata = meta;
 		++meta;
-		world.setBlockMetadataWithNotify(x, y, z, meta, 3);
+		world.setBlockMetadataWithNotify(x, y, z, meta, 2);
 		AppleCore.announceGrowthTick(this, world, x, y, z, previousMetadata);
 	}
 
@@ -66,30 +61,26 @@ public class BlockGrapeVine1 extends Block implements IPlantable
 		// 2 - hard sapling
 		// 3 - trunk
 
-		Event.Result allowGrowthResult = AppleCore.validateGrowthTick(this, world, x, y, z, random);
-		if (allowGrowthResult == Event.Result.DENY)
-			return;
-
-		int meta = world.getBlockMetadata(x, y, z);
-		float f = this.getGrowthRate(world, x, y, z);
-
-		boolean continueGrowth = random.nextInt((int)(this.growth / f) + 1) == 0;
-		if (allowGrowthResult == Event.Result.ALLOW || continueGrowth)
+		if (world.getBlockLightValue(x, y + 1, z) >= 9)
 		{
-			/* Is here a rope block above this? */
-			if (meta == 0 && world.getBlock(x, y + 1, z) == GrowthCraftCore.ropeBlock)
+			Event.Result allowGrowthResult = AppleCore.validateGrowthTick(this, world, x, y, z, random);
+			if (allowGrowthResult == Event.Result.DENY)
+				return;
+
+			int meta = world.getBlockMetadata(x, y, z);
+			float f = this.getGrowthRate(world, x, y, z);
+
+			boolean continueGrowth = random.nextInt((int)(this.growth / f) + 1) == 0;
+			if (allowGrowthResult == Event.Result.ALLOW || continueGrowth)
 			{
-				incrementGrowth(world, x, y, z, meta);
-				world.setBlock(x, y + 1, z, GrowthCraftGrapes.grapeLeaves, 0, 3);
-			}
-			if (meta == 0 && world.isAirBlock(x, y + 1, z))
-			{
-				incrementGrowth(world, x, y, z, meta);
-				world.setBlock(x, y + 1, z, this, 0, 3);
-			}
-			else if (meta == 0 && world.getBlock(x, y + 1, z) == GrowthCraftGrapes.grapeLeaves)
-			{
-				incrementGrowth(world, x, y, z, meta);
+				if (meta == 0)
+				{
+					incrementGrowth(world, x, y, z, meta);
+				}
+				else if (meta == 1)
+				{
+					world.setBlock(x, y, z, GrowthCraftGrapes.grapeVine1, 0, 3);
+				}
 			}
 		}
 	}
@@ -102,19 +93,14 @@ public class BlockGrapeVine1 extends Block implements IPlantable
 	private float getGrowthRate(World world, int x, int y, int z)
 	{
 		float f = 1.0F;
-		int j = y;
-		if (world.getBlock(x, j - 1, z) == this && world.getBlock(x, j - 2, z) == Blocks.farmland)
-		{
-			j = y - 1;
-		}
-		Block l = world.getBlock(x, j, z - 1);
-		Block i1 = world.getBlock(x, j, z + 1);
-		Block j1 = world.getBlock(x - 1, j, z);
-		Block k1 = world.getBlock(x + 1, j, z);
-		Block l1 = world.getBlock(x - 1, j, z - 1);
-		Block i2 = world.getBlock(x + 1, j, z - 1);
-		Block j2 = world.getBlock(x + 1, j, z + 1);
-		Block k2 = world.getBlock(x - 1, j, z + 1);
+		Block l = world.getBlock(x, y, z - 1);
+		Block i1 = world.getBlock(x, y, z + 1);
+		Block j1 = world.getBlock(x - 1, y, z);
+		Block k1 = world.getBlock(x + 1, y, z);
+		Block l1 = world.getBlock(x - 1, y, z - 1);
+		Block i2 = world.getBlock(x + 1, y, z - 1);
+		Block j2 = world.getBlock(x + 1, y, z + 1);
+		Block k2 = world.getBlock(x - 1, y, z + 1);
 		boolean flag = this.isGrapeVine(j1) || this.isGrapeVine(k1);
 		boolean flag1 = this.isGrapeVine(l) || this.isGrapeVine(i1);
 		boolean flag2 = this.isGrapeVine(l1) || this.isGrapeVine(i2) || this.isGrapeVine(j2) || this.isGrapeVine(k2);
@@ -123,14 +109,14 @@ public class BlockGrapeVine1 extends Block implements IPlantable
 		{
 			for (int i3 = z - 1; i3 <= z + 1; ++i3)
 			{
-				Block block = world.getBlock(l2, j - 1, i3);
+				Block block = world.getBlock(l2, y - 1, i3);
 				float f1 = 0.0F;
 
 				if (block != null && block == Blocks.farmland)
 				{
 					f1 = 1.0F;
 
-					if (block.isFertile(world, l2, j - 1, i3))
+					if (block.isFertile(world, l2, y - 1, i3))
 					{
 						f1 = 3.0F;
 					}
@@ -162,7 +148,7 @@ public class BlockGrapeVine1 extends Block implements IPlantable
 		if (!this.canBlockStay(world, x, y, z))
 		{
 			this.dropBlockAsItem(world, x, y, z, world.getBlockMetadata(x, y, z), 0);
-			world.setBlock(x, y, z, Blocks.air, 0, 2);
+			world.setBlock(x, y, z, Blocks.air, 0, 3);
 		}
 	}
 
@@ -173,7 +159,7 @@ public class BlockGrapeVine1 extends Block implements IPlantable
 	public boolean canBlockStay(World world, int x, int y, int z)
 	{
 		Block soil = world.getBlock(x, y - 1, z);
-		return (soil != null && soil.canSustainPlant(world, x, y - 1, z, ForgeDirection.UP, this)) || world.getBlock(x, y - 1, z) == this;
+		return soil != null && soil.canSustainPlant(world, x, y - 1, z, ForgeDirection.UP, this);
 	}
 
 	/************
@@ -198,13 +184,13 @@ public class BlockGrapeVine1 extends Block implements IPlantable
 	@Override
 	public Item getItemDropped(int meta, Random par2Random, int par3)
 	{
-		return GrowthCraftGrapes.grapeSeeds;
+		return null;
 	}
 
 	@Override
 	public int quantityDropped(Random random)
 	{
-		return 1;
+		return 0;
 	}
 
 	/************
@@ -214,24 +200,17 @@ public class BlockGrapeVine1 extends Block implements IPlantable
 	@SideOnly(Side.CLIENT)
 	public void registerBlockIcons(IIconRegister reg)
 	{
-		tex = new IIcon[3];
+		tex = new IIcon[2];
 
-		tex[0] = reg.registerIcon("grcgrapes:trunk");
-		tex[1] = reg.registerIcon("grcgrapes:leaves");
-		tex[2] = reg.registerIcon("grcgrapes:leaves_opaque");
+		tex[0] = reg.registerIcon("grcgrapes:vine_0");
+		tex[1] = reg.registerIcon("grcgrapes:vine_1");
 	}
 
 	@Override
 	@SideOnly(Side.CLIENT)
 	public IIcon getIcon(int side, int meta)
 	{
-		return this.tex[0];
-	}
-
-	public IIcon getLeafTexture()
-	{
-		this.graphicFlag = ((BlockLeaves)Blocks.leaves).isOpaqueCube();
-		return !this.graphicFlag ? this.tex[1] : this.tex[2];
+		return this.tex[meta];
 	}
 
 	/************
@@ -240,7 +219,7 @@ public class BlockGrapeVine1 extends Block implements IPlantable
 	@Override
 	public int getRenderType()
 	{
-		return RenderGrapeVine1.id;
+		return 1;
 	}
 
 	@Override
@@ -259,25 +238,9 @@ public class BlockGrapeVine1 extends Block implements IPlantable
 	 * BOXES
 	 ************/
 	@Override
-	public void addCollisionBoxesToList(World world, int x, int y, int z, AxisAlignedBB aabb, List list, Entity entity)
+	public AxisAlignedBB getCollisionBoundingBoxFromPool(World world, int x, int y, int z)
 	{
-		int meta = world.getBlockMetadata(x, y, z);
-		float f = 0.0625F;
-
-		if (meta == 0)
-		{
-			this.setBlockBounds(6*f, 0.0F, 6*f, 10*f, 0.5F, 10*f);
-			super.addCollisionBoxesToList(world, x, y, z, aabb, list, entity);
-			this.setBlockBounds(4*f, 0.5F, 4*f, 12*f, 1.0F, 12*f);
-			super.addCollisionBoxesToList(world, x, y, z, aabb, list, entity);
-		}
-		else if (meta == 1)
-		{
-			this.setBlockBounds(6*f, 0.0F, 6*f, 10*f, 1.0F, 10*f);
-			super.addCollisionBoxesToList(world, x, y, z, aabb, list, entity);
-		}
-
-		this.setBlockBoundsBasedOnState(world, x, y, z);
+		return null;
 	}
 
 	@Override
@@ -289,10 +252,10 @@ public class BlockGrapeVine1 extends Block implements IPlantable
 		switch (meta)
 		{
 		case 0:
-			this.setBlockBounds(4*f, 0.0F, 4*f, 12*f, 1.0F, 12*f);
+			this.setBlockBounds(6*f, 0.0F, 6*f, 10*f, 5*f, 10*f);
 			break;
 		case 1:
-			this.setBlockBounds(6*f, 0.0F, 6*f, 10*f, 1.0F, 10*f);
+			this.setBlockBounds(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F);
 			break;
 		}
 	}
