+++ src/java/growthcraft/grapes/renderer/RenderGrape.java
@@ -1,13 +1,15 @@
-package growthcraft.grapes;
+package growthcraft.grapes.renderer;
 
 import growthcraft.core.Utils;
+import growthcraft.grapes.block.BlockGrapeBlock;
+
+import cpw.mods.fml.client.registry.ISimpleBlockRenderingHandler;
+import cpw.mods.fml.client.registry.RenderingRegistry;
 import net.minecraft.block.Block;
 import net.minecraft.client.renderer.RenderBlocks;
 import net.minecraft.client.renderer.Tessellator;
 import net.minecraft.util.IIcon;
 import net.minecraft.world.IBlockAccess;
-import cpw.mods.fml.client.registry.ISimpleBlockRenderingHandler;
-import cpw.mods.fml.client.registry.RenderingRegistry;
 
 public class RenderGrape implements ISimpleBlockRenderingHandler
 {
