+++ src/java/growthcraft/grapes/renderer/RenderGrapeLeaves.java
@@ -1,6 +1,10 @@
-package growthcraft.grapes;
+package growthcraft.grapes.renderer;
 
 import growthcraft.core.Utils;
+import growthcraft.grapes.block.BlockGrapeLeaves;
+
+import cpw.mods.fml.client.registry.ISimpleBlockRenderingHandler;
+import cpw.mods.fml.client.registry.RenderingRegistry;
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockLeaves;
 import net.minecraft.client.renderer.RenderBlocks;
@@ -8,11 +12,7 @@ import net.minecraft.client.renderer.Tessellator;
 import net.minecraft.init.Blocks;
 import net.minecraft.util.IIcon;
 import net.minecraft.world.IBlockAccess;
-
 import org.lwjgl.opengl.GL11;
-
-import cpw.mods.fml.client.registry.ISimpleBlockRenderingHandler;
-import cpw.mods.fml.client.registry.RenderingRegistry;
 
 public class RenderGrapeLeaves implements ISimpleBlockRenderingHandler
 {
