+++ src/java/growthcraft/grapes/renderer/RenderGrapeVine1.java
@@ -1,16 +1,16 @@
-package growthcraft.grapes;
+package growthcraft.grapes.renderer;
 
 import growthcraft.core.Utils;
+import growthcraft.grapes.block.BlockGrapeVine1;
+
+import cpw.mods.fml.client.registry.ISimpleBlockRenderingHandler;
+import cpw.mods.fml.client.registry.RenderingRegistry;
 import net.minecraft.block.Block;
 import net.minecraft.client.renderer.RenderBlocks;
 import net.minecraft.client.renderer.Tessellator;
 import net.minecraft.world.ColorizerFoliage;
 import net.minecraft.world.IBlockAccess;
-
 import org.lwjgl.opengl.GL11;
-
-import cpw.mods.fml.client.registry.ISimpleBlockRenderingHandler;
-import cpw.mods.fml.client.registry.RenderingRegistry;
 
 public class RenderGrapeVine1 implements ISimpleBlockRenderingHandler
 {
